Setup a pdfNode stresser
**Motivation:**
Having the new pdfNode production/staging machines up and running we
want a way (wihtout using the malaca's stressing machine) to call these
pdfNodes and load test them.

**Modifications:**
We neede the documents data examples to generate the pdfs. On the docs
folder were created one invoice small (having +50 items), a medium one
(having +100 items) and a big one (with +150 items).
The project.clj was updated about some libraries needed and the
stresser was created on the ix_stresser folder and called: pdfnode.clj.

**Result:**
From now one, to load stress pdfNode about the generation of pdfs, we
will not need the malaca's stressing machine and we make Marcos proud.
We have a SCRIPT! :D