Create pdfNode bulk stresser
**Motivation:**
Just has we want to stress pdf generators endpoints of pdfNode, we want
too to stress the endpoints about bulks. We want and need to test all
these endpoints and check pdfNode reaction on big loads.

**Modifications:**
A new file was created, pdfbulknode.clj. This one, just like
pdfnode.clj, has the code and the request by default the
`applyTemplateBulk` endpoint but we can pass as parameter the endpoint
we want to stress out.

**Result:**
The stresser is looking good. Some tests were made and pdfNode responde
as expected so let the games begin and try to break pdfNode :P