Improve pdfbulknode
**Motivation:**
While using it on load tests i saw a lot of erros that mainly I think
they were because of having writting and reading from the same folder
and files.

**Modifications:**
To solf this issue the solution was pretty simple, just change the
filename of the bulk, so the folders have different names.
Plus, there are some small modifications to clean more the code.

**Result:**
Each bulk request now will have a different filename, wich means on the
end, different folder where the pdfs are created, saved and zipped.