Get sensitive data secured
**Motivation:**
Lets do not use tokens, urls, or any sensitive data about our staging
or production environments directly hard coded on the code or on the
sample docs.

**Modifications:**
Start to use profiles for the pdfnode and pdfbulknode stressers.
The pdfNode token used for the requests and staging/production
respective urls are now safe on the profiles.clj.

**Result:**
The resquest are looking lood. The data is there ;)